refactor: move interfaces to special folder and add comments
