+++ src/interfaces/standard.ts
@@ -1,0 +1,6 @@
+import {ReactNode} from 'react'
+
+// Standard components props
+export interface ChildrenProps {
+  children: ReactNode
+}
